Allow selection of paths wider than Emacs window
Previously, when fzf finished and printed to the terminal buffer a path
that was wider than the Emacs window, term mode was adding a newline at
each place where the path reached the right edge of the window and
needed to wrap to the next line.  With this change, term mode does not
add any newlines, so fzf can safely deliver long paths back to Emacs.