Fix use of setq-local
The setq-local macro accepts only one pair of parameters until it was changed
in Emacs 27 to support multiple pairs.  Since the file states it requires emacs 24.4,
then the code must use the form that was available then and which is still valid.
There is no performance issue in this code to justify using the newer form.