Issue #84 Fix.  Also fixes issue #82.
Fixes:

- https://github.com/bling/fzf.el/issues/84
- https://github.com/bling/fzf.el/issues/82

'fzf/after-term-handle-exit' now detect extra spaces or other characters
at the end of the file name returned by extraction of fzf output.

This occurs since fzf 0.35.1 as reported by issue #82 because of a new
separator line.  But instead of modifying the command line the function trims
any whitespace at the end of the file name.  If the resulting file name is not
found it tries to extract the correct file name by removing characters at the
end until a valid filename is found.  If it does not find any valid file name
it then issues a descriptive message to provide enough info for further
improvements.