Add fzf-find-in-buffer (by LiuYinCarl) (#96)
* Describe dwim with grep limitation. +1 optimization.

* Add fzf-find-in-buffer (by LiuYinCarl)

Add fzf-find-in-buffer originally proposed by LiuYinCarl in
https://github.com/bling/fzf.el/issues/94

I modified the proposed code a little to adapt to the current code and added
protection against trying to search in a buffer that is not visiting a file,
an prompt to save the buffer if it has been modified.

* Fix issue #95: dwim with grep

A work-around fix for dwim operations using grep.  When a FZF grep with
specified text is executed, a file pattern of " *" is automatically appended
to the grep command, allowing the operation to proceed.