Fix narrowing grep fuzzy searches.  More flexible narrowing commands.
This addresses issue #98: it allows narrowing fzf searches to work with grep.

It also adds the ability to invoke the grep narrowing commands with a prefix
argument to get prompted for a file pattern.  The file pattern for grep and
Ripgrep differ, so the prompt identifies the tool used (grep or rg) if it
recognizes it, otherwise it just prompts for a file pattern.

Committing with commented-out tracing message calls I will remove in next
commit; to keep in the history for reference.