More flexible dwim grep with prompt history (#97)
* More flexible dwim grep with prompt history

The `fzf-grep-dwim` and `fzf-grep-dwim-with-narrowing` commands can now be
invoked with a command prefix, like C-u.  When that's done the commands prompt
for the file pattern to use.  Otherwise they use the default file pattern
identified by the new `fzf/grep-file-pattern` customizable user-option.

All prompts for grep operations now have a prompt history which can be invoked
during the minibuffer prompt operation with M-p and M-p.  By default the
history is kept independent for each major mode. If you prefer a global prompt
history, shared across all major modes, then set the new
`fzf/prompt-history-per-major-mode` customizable user-option to nil.

Removed the limitation notes in the grep dwim commands as they no longer
apply.

Add :safe attribute to the boolean user-options.

* Complete docstring.  Fix fzf-grep-dwim on no symbol.

* Fix narrowing grep fuzzy searches.  More flexible narrowing commands.

This addresses issue #98: it allows narrowing fzf searches to work with grep.

It also adds the ability to invoke the grep narrowing commands with a prefix
argument to get prompted for a file pattern.  The file pattern for grep and
Ripgrep differ, so the prompt identifies the tool used (grep or rg) if it
recognizes it, otherwise it just prompts for a file pattern.

Committing with commented-out tracing message calls I will remove in next
commit; to keep in the history for reference.

* Remove debug tracing message calls

* Enhanced narrowing command docstrings.