Fix issue #99: allow using color for Emacs in terminal
This addresses https://github.com/bling/fzf.el/issues/99.

When Emacs is not running in graphics mode the new `fzf--text-only-for`
function now used by `fzf--grep-cmd` and `fzf--after-term-handle-exit` to
extract the text now filter the ANSI color code sequences when necessary, and
also filter the character properties (as only done before).