More informative message on FZF selection.
The current message simply shows "FZF exited with code 0" when it works. This
message might have been useful during debugging but it's not very useful for
users.

On success, I replaced it with a description of what was actually selected; a
file name (with full path), file name, line number and line text, or the line
number and line text depending of the type of command executed.

On error, a user-error showing the FZF error message is shown.

The message is most useful when a user wants to remember what file was
previously selected by the FZF operation.