fzf-switch-buffer: default to most-recently-used other buffer
Previously, the default buffer for `fzf-switch-buffer` was the *current*
buffer, which doesn't seem like a useful thing to "switch" to.

Vanilla `switch-buffer` has the useful property that a bare `RET`
response switches to the most-recently-used-before-the-current-buffer
buffer. This commit makes `fzf-switch-buffer` behave the same way by
deleting the current buffer from the entries list.